refactor: house NodeUi response types in response.rs; rename ValidateCommitted
- Move the `NodeUi` method response types (`NodeUiResponse`,
  `InspectorRowsResponse`, `InspectorUiResponse`, `ContextMenuResponse`) and
  their shared builder macro from `gantz_egui` lib root into the `response`
  module (alongside `Responses`/`DynResponse`), re-exporting them from the
  crate root so existing paths are unchanged.
- Rename the `bevy_gantz::vm::ChangeTrackingValidation` resource to
  `ValidateCommitted` to match its `validate_committed` system.
